+++ CHANGELOG.md
@@ -10,13 +10,13 @@
   * Fixed issue where an error was thrown when no DeveloperDashboard
     parameter was specfied
 * SPInstall
-  * Added check to unblock setup file it is blocked because it is coming
+  * Added check to unblock setup file if it is blocked because it is coming
     from a network location. This to prevent endless wait
   * Added ability to install from a UNC path, by adding server
     to IE Local Intranet Zone. This will prevent an endless wait
     caused by security warning
 * SPInstallLanguagePack
-  * Added check to unblock setup file it is blocked because it is coming
+  * Added check to unblock setup file if it is blocked because it is coming
     from a network location. This to prevent endless wait
   * Corrected issue with Norwegian language pack not being correctly
     detected